I present, *THE* *CORE*, with the "to_ascii" function". Eventually much of what is in Products.XWFCore.XWFUtils will be moved to gs.core, but that is for another day. Ensuring that the TextMixin.set_headers method behaves itself in gs.content.email.base Switching to Unicode literals, and calling gs.core.to_ascii when setting the headers in:   gs.content.form.api.json   gs.errormesg   gs.profile.image.base   gs.profile.image.square   gs.profile.invite
Switching to Unicode literals, calling gs.core.to_ascii when setting the header, switching the base-class to TextMixin, and updating the Welcome message so it uses the standard email-layout in gs.group.member.add.base
Switching to Unicode literals, calling gs.core.to_ascii, and switching the base-class to TextMixin in:
  gs.group.member.canpost

Switching to Unicode literals, calling gs.core.to_ascii.
  gs.group.member.invite.base

Switching to Unicode literals, calling gs.core.to_ascii, and switching the base-class to TextMixin:
  gs.group.member.join
  gs.group.member.request

Switching to Unicode literals, and calling gs.core.to_ascii when setting the header (in the error-pages):
  gs.group.messages.image

Switching to Unicode literals, and calling gs.core.to_ascii when setting the header (when returning JSON objects):
  gs.group.messages.starttopic

Switching to Unicode literals, and calling gs.core.to_ascii when setting the header (when returning values from the Sticky Topic getter):
  gs.group.messages.topic

Switching to Unicode literals, and calling gs.core.to_ascii when setting the headers:
  gs.group.messages.topicsdigest
  gs.group.start

Switching to Unicode literals, and calling gs.core.to_ascii when setting the headers. (The reason that gs.login needs to set a header is still a mystery.):
  gs.login

Switching to Unicode literals, calling gs.core.to_ascii when setting the headers (on the error pages), and switching the base-class to TextMixin:
 gs.profile.email.verify

A Unicode and metata fix for the audit-trail system:
  Products.GSAuditTrail

An update to the egg metadata:
  gs.viewlet